fix preview mode

+++ os/arch/polybar/launch.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
-killall -q polybar
+if [ -z $PREVIEW ]; then
+    killall -q polybar
 
-while pgrep -x polybar >/dev/null; do sleep 1; done
+    while pgrep -x polybar >/dev/null; do sleep 1; done
+fi
 
 polybar top &
